+++ src/resources/containers/index.ts
@@ -3,8 +3,9 @@ import * as Instances from "./instances";
 import * as CompatibleImages from "./compatible-images";
 import * as Telemetry from "./telemetry";
 import * as Config from "./config";
+import * as Backups from "./backups";
 
-export { Servers, Instances, CompatibleImages, Telemetry, Config };
+export { Servers, Instances, CompatibleImages, Telemetry, Config, Backups };
 
 export * from "./container";
 export * from "./tasks";