+++ src/resources/stacks/spec/v1/integrations.ts
@@ -4,6 +4,7 @@ export interface Integrations {
   webhooks?: Webhooks;
   lets_encrypt?: LetsEncrypt;
   files?: File[];
+  backups?: Backups;
 }
 
 export interface LetsEncrypt {
@@ -29,4 +30,23 @@ export interface Events {
 export interface File {
   source: string;
   destination: string;
+}
+
+export interface Backups {
+  destination: BackupDestination;
+  backup: Backup;
+  restore: Restore | null;
+}
+
+export type BackupDestination = "backblaze-b2";
+
+export interface Backup {
+  command: string;
+  timeout: number | null;
+  cron_string: string | null;
+}
+
+export interface Restore {
+  command: string;
+  timeout: number | null;
 }